Re-word README
Hiya!  Hopefully you will not take offence to this PR as it is merely cleaning up the English in the README.  I'm happy to revert whatever you'd like, discuss anything, and of course, I'm cool if this PR is just outright rejected.

My aim was to reword it to be more "idiomatic English" without changing the meaning of what you are saying.  In general, I'm just excited about this project—I would love to have `defq` in my life—and this was a small first step I saw to contribute.